Update company middleware to TS
